combining app_context_render and app_root_provider
